Part one solved. Found error in digits() method, long int instead of long long.

+++ Day2/day2.cpp
@@ -1,7 +1,7 @@
 /////////////////////////////////////////////////
 //
 // Advent of Code 2025
-// Day 2 - 
+// Day 2 - Part one and two
 // Author: Fritz, M.H.
 //
 //
@@ -13,7 +13,7 @@
 #include <string>
 using namespace std;
 
-int digits(long int id){
+int digits(long long id){
     //Tells how many digits an id has
     int digits = 0;
     while (id) {
@@ -28,7 +28,7 @@ long long isIDvalid(long long id){
     //Checks if ID is valid.
     //If yes, return id
     //If no, return 0
-        
+    
     int numDig = digits(id);
     if(numDig%2 != 0) return 0; //exclude all IDs with odd number of digits
     
@@ -70,7 +70,7 @@ int main() {
         
         nextRange = input.substr(nextChar,commaPos-nextChar);
         cout << "\n(" + to_string(commaPos) + ") " + nextRange + " ";
-        
+
         rangeStart = stoll(nextRange.substr(0, nextRange.find("-")));
         rangeEnd = stoll(nextRange.substr(nextRange.find("-") + 1, nextRange.size() - nextRange.find("-")));
         for(long long id = rangeStart; id <= rangeEnd; id++){
@@ -79,7 +79,6 @@ int main() {
                 cout << "\ninvalid id:" + to_string(id) + " new answer:" + to_string(answer);
             } 
         }
-        //cout << "new answer: " + to_string(answer) + "\n";
 
         nextChar = commaPos + 1;
         if(commaPos == std::string::npos){
